+++ src/utils/stopPropagation.js
@@ -1,0 +1,8 @@
+import curry from 'lodash/fp/curry'
+
+const stopPropagation = curry((callback, evt, ...args) => {
+  evt.stopPropagation()
+  callback(evt, ...args)
+})
+
+export default stopPropagation